// src/components/three/Common/interfaces.tsx
export interface IRotaion {
  rotaionX: number;
  rotaionY: number;
  rotaionZ: number;
  anisotropy?: number;
  opacity: number;
  args: any;
}

export interface IClickablePoints {
  x: number;
  y: number;
  z: number;
  tootltipTitle: string;
  tootltipDescription: string;
}

export const ConstSphereBackgroundImg = "https://i.ibb.co/5kFytqr/bg2.jpg";